dsp: make the float-to-double promotions explicit
Thirteen -Wdouble-promotion warnings in the strict build, all of the same shape:
a float multiplied by a double, in a place where computing in double is the
intent. The conversion the compiler was already performing IS static_cast<double>,
so writing it out changes nothing except that the reader can see the decision.

kTwoPi deliberately stays a float and is cast rather than replaced with a double
2*pi. The more accurate constant would move every partial's rotation, and the
fingerprint would report that as a regression it is not.

Verified rather than assumed: the fingerprint is byte-identical across the
change.

+++ src/caecilia/synthesis/ArEnvelope.h
@@ -31,8 +31,8 @@ public:
         const double sr = sampleRate > 1.0 ? sampleRate : 48000.0;
         const float a = attackSeconds  > 0.0f ? attackSeconds  : 0.0005f;
         const float r = releaseSeconds > 0.0f ? releaseSeconds : 0.0005f;
-        attackStep_  = static_cast<float>(1.0 / (a * sr));
-        releaseStep_ = static_cast<float>(1.0 / (r * sr));
+        attackStep_  = static_cast<float>(1.0 / (static_cast<double>(a) * sr));
+        releaseStep_ = static_cast<float>(1.0 / (static_cast<double>(r) * sr));
     }
 
     /// Begin the attack from the current gain (retrigger-safe). RT-safe.
@@ -50,7 +50,7 @@ public:
     {
         const double sr = sampleRate > 1.0 ? sampleRate : 48000.0;
         const float r = releaseSeconds > 0.0f ? releaseSeconds : 0.0005f;
-        releaseStep_ = static_cast<float>(1.0 / (r * sr));
+        releaseStep_ = static_cast<float>(1.0 / (static_cast<double>(r) * sr));
     }
 
     /// Advance one sample and return the current gain in [0, 1]. RT-safe.
